requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
filelock==3.18.0
flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
pillow==10.4.0
proglog==0.1.11
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sympy==1.13.3
tiktoken==0.9.0
torch==2.2.2
torchaudio==2.2.2
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
zipp==3.21.0
